link: add portal error for composite case
